llamaIndex-example/requirements.txt
# 核心框架
langchain==0.1.0
llama-index==0.9.30

# 向量数据库
pymilvus==2.3.4

# 本地模型支持
ollama==0.1.7

# 数据处理
numpy==1.24.3
pandas==2.0.3
pillow==10.0.1

# API框架
fastapi==0.104.1
uvicorn==0.24.0

# 环境配置
python-dotenv==1.0.0
pydantic==2.5.0

# 日志和监控
loguru==0.7.2

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 其他工具
requests==2.31.0
aiofiles==23.2.1
typing-extensions==4.8.0